ISSUE #611 refactor units to its own file, use synchro 4.0.1
